Divide app.py into multiple .py files
