+++ hostcomp/models.py
@@ -1,0 +1,15 @@
+from . import db
+
+
+class Score(db.Model):
+    __tablename__ = 'scores'
+    name = db.Column(db.String(), primary_key=True)
+    public = db.Column(db.Float, nullable=False)
+    private = db.Column(db.Float, nullable=False)
+
+
+class Competition(db.Model):
+    __tablename__ = 'competitions'
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(), nullable=False)
+    disclose_private = db.Column(db.Boolean, nullable=False)
