+++ hostcomp/templates/admin.html
